fix: adding validations

+++ src/domain/use-cases/tablero/delete.use-case.ts
@@ -1,4 +1,5 @@
 import {TableroEntity} from '../../entities';
+import {CustomError} from '../../errors/CustomError';
 import {TableroRepository} from '../../repository';
 
 export interface DeleteTableroUseCase {
@@ -8,6 +9,10 @@ export interface DeleteTableroUseCase {
 export class DeleteTablero implements DeleteTableroUseCase {
 	constructor(private readonly repository: TableroRepository) {}
 	async execute(id: number): Promise<TableroEntity> {
+		const tablero = await this.repository.findById(id);
+		if (!tablero) {
+			throw CustomError.NotFound('No se encontró registro');
+		}
 		return this.repository.deleteById(id);
 	}
 }
